Fix typos in `stores.rs`
Due to four typos in the match block starting at line 30, caused by
GitHub Copilot attempting to format the URLs to the images in the same
way as the "shopify" case and assuming that multiple URLs like
`https://github.com/subreme/{storename}-monitor` existed, the program
would generate invalid links to the logos of "atmos", "Concepts", "Extra
Butter", and "HANON", leading to webhook embeds with empty author
images. This issue could have been avoided by checking the code
generated by the tool, and was most likely missed due to previous tests
being unable to demonstrate it as the repository was not public yet, and
therefore the images weren't either.