Fix persistent settings
Due to how the logic that checked settings at every "level" before
obtaining the correct values for each event was written with a single
variable per individual setting, if a configuration was set to a
non-Default value for an event, and the successive ones didn't include
it, they would also share the same value. This was fixed by cloning each
value at every "level" of the recursion and resetting the values to the
cloned "higher level" ones at every iteration.

While looking for the cause of the bug, some other related bugs were
fixed, such as the wrongful recognition of `null` as a missing field in
the `struct`s where the two values have different meanings, due to how the
fields were defined in `config.rs`.

Additionally, several instances of `.into()` being called in places
where `.clone()` would have been more idiomatic were changed to the
latter method.