Add `minimum` field to `Settings`
Added a `minimum` field to the `Settings` `struct` in `config.rs`. This
is used to control the minimum number of available variants a product
should have for a webhook to be sent out.

While it is already fully functional, its default value of 0 will not
change how the monitor works if `config.json` is not updated.

Once the feature has gone through proper testing and more commits are
pushed, it will be referenced in `README.md` and the project's version
will be bumped.