Fix Clippy's `too_many_arguments` lint
Removed `#![allow(clippy::too_many_arguments)]` from `monitor.rs`,
creating the `ItemSettings` and `PasswordSettings` structs to be used as
the single parameters for `item()` and `password()` respectively.

Modified the `hidden!()` macro to print its messages to the console when
the `dev` profile is used (when `debug` mode is active), without
affecting its functionality in release mode.